Rename inspector to inspection_util
