+++ modi/__main__.py
@@ -10,7 +10,7 @@ import modi
 from modi.util.debugging_util import Debugger
 from modi.util.usage_instructor import UsageInstructor
 from modi.util.tutor import Tutor
-from modi.util.inspector import Inspector
+from modi.util.inspection_util import Inspector
 
 from modi.util.firmware_updater import STM32FirmwareUpdater
 from modi.util.firmware_updater import ESP32FirmwareUpdater
